hook up save height/weight/bday

+++ server/migrations/000011_add_cols_weight_height_birthday_calories.up.sql
@@ -2,8 +2,8 @@ BEGIN;
 
 ALTER TABLE users
 ADD COLUMN IF NOT EXISTS birthdate timestamp with time zone,
-ADD COLUMN IF NOT EXISTS weight integer,
-ADD COLUMN IF NOT EXISTS height integer,
+ADD COLUMN IF NOT EXISTS weight numeric,
+ADD COLUMN IF NOT EXISTS height numeric,
 ADD COLUMN IF NOT EXISTS is_male boolean;
 
 ALTER TABLE metrics_top_level
